Better sorting for conferences
This refines the sorting for conferences. Now when two upcoming conferences start on the same date, the longer conference will be listed first. If two conferences have the same start and end dates, they will be listed alphabetically.

For past conferences, if two conferences have the same end date, the shorter conference will be listed first (because it started sooner). If two past conferences have the same start and end dates, they will be listed alphabetically.